Moved Proof of Concept (POC) logic from HomeFragment.java to YelpService.java
