--- app/src/main/java/com/example/whereto/models/YelpFusion.java
@@ -1,32 +1,0 @@
-package com.example.whereto.models;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.Properties;
-
-public class YelpFusion {
-
-    @SerializedName("type")
-    @Expose
-    private String type;
-    @SerializedName("properties")
-    @Expose
-    private Properties properties;
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public Properties getProperties() {
-        return properties;
-    }
-
-    public void setProperties(Properties properties) {
-        this.properties = properties;
-    }
-}
